@@ -1,35 +1,27 @@
-﻿namespace CleanArchitecture.Common.Behaviours
+﻿namespace CleanArchitecture.Common.Behaviours;
+
+public class ValidationBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
+    where TRequest : IRequest<TResponse>
 {
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Threading;
-    using System.Threading.Tasks;
-    using FluentValidation;
-    using MediatR;
+    private readonly IEnumerable<IValidator<TRequest>> validators;
 
-    public class ValidationBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
-        where TRequest : IRequest<TResponse>
+    public ValidationBehavior(IEnumerable<IValidator<TRequest>> validators) => this.validators = validators;
+
+    public async Task<TResponse> Handle(TRequest request, RequestHandlerDelegate<TResponse> next, CancellationToken cancellationToken)
     {
-        private readonly IEnumerable<IValidator<TRequest>> validators;
-
-        public ValidationBehavior(IEnumerable<IValidator<TRequest>> validators) => this.validators = validators;
-
-        public async Task<TResponse> Handle(TRequest request, CancellationToken cancellationToken, RequestHandlerDelegate<TResponse> next)
+        if (this.validators.Any())
         {
-            if (this.validators.Any())
+            var context = new ValidationContext<TRequest>(request);
+
+            var validationResults = await Task.WhenAll(this.validators.Select(v => v.ValidateAsync(context, cancellationToken)));
+            var failures = validationResults.SelectMany(r => r.Errors).Where(f => f != null);
+
+            if (failures.Any())
             {
-                var context = new ValidationContext<TRequest>(request);
-
-                var validationResults = await Task.WhenAll(this.validators.Select(v => v.ValidateAsync(context, cancellationToken)));
-                var failures = validationResults.SelectMany(r => r.Errors).Where(f => f != null);
-
-                if (!failures.Any())
-                {
-                    throw new ValidationException(failures);
-                }
+                throw new ValidationException(failures);
             }
-
-            return await next();
         }
+
+        return await next();
     }
 }
