@@ -1,27 +1,24 @@
-﻿namespace CleanArchitecture.Common.Exceptions
+﻿namespace CleanArchitecture.Common.Exceptions;
+
+public class NotFoundException : Exception
 {
-    using System;
-
-    public class NotFoundException : Exception
+    public NotFoundException()
+        : base()
     {
-        public NotFoundException()
-            : base()
-        {
-        }
+    }
 
-        public NotFoundException(string message)
-            : base(message)
-        {
-        }
+    public NotFoundException(string message)
+        : base(message)
+    {
+    }
 
-        public NotFoundException(string message, Exception innerException)
-            : base(message, innerException)
-        {
-        }
+    public NotFoundException(string message, Exception innerException)
+        : base(message, innerException)
+    {
+    }
 
-        public NotFoundException(string name, object key)
-            : base($"Entity \"{name}\" ({key}) was not found.")
-        {
-        }
+    public NotFoundException(string name, object key)
+        : base($"Entity \"{name}\" ({key}) was not found.")
+    {
     }
 }
