@@ -1,31 +1,25 @@
-﻿namespace CleanArchitecture.Common.Exceptions
+﻿namespace CleanArchitecture.Common.Exceptions;
+
+public class ValidationException : Exception
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using FluentValidation.Results;
+    public ValidationException()
+        : base("One or more validation failures have occurred.")
+        => this.Errors = new Dictionary<string, string[]>();
 
-    public class ValidationException : Exception
+    public ValidationException(IEnumerable<ValidationFailure> failures)
+        : this()
     {
-        public ValidationException()
-            : base("One or more validation failures have occurred.")
-            => this.Errors = new Dictionary<string, string[]>();
+        var failureGroups = failures
+            .GroupBy(e => e.PropertyName, e => e.ErrorMessage);
 
-        public ValidationException(IEnumerable<ValidationFailure> failures)
-            : this()
+        foreach (var failureGroup in failureGroups)
         {
-            var failureGroups = failures
-                .GroupBy(e => e.PropertyName, e => e.ErrorMessage);
+            var propertyName = failureGroup.Key;
+            var propertyFailures = failureGroup.ToArray();
 
-            foreach (var failureGroup in failureGroups)
-            {
-                var propertyName = failureGroup.Key;
-                var propertyFailures = failureGroup.ToArray();
-
-                this.Errors.Add(propertyName, propertyFailures);
-            }
+            this.Errors.Add(propertyName, propertyFailures);
         }
-
-        public IDictionary<string, string[]> Errors { get; }
     }
+
+    public IDictionary<string, string[]> Errors { get; }
 }
